Update to 1.12.1
- Remove built in packaging of cofh RF api, and switch to optional usage of RedstoneFlux library mod
- Disabled MCMultipart support as their API changed a lot and the only mod that depends on it is PL2
- Fixed Mekanism Integration (Mekanism will cause a crash if RedstoneFlux is not installed until they next update)
- Includes all of the patches to the 1.10.2 version of SonarCore that was not ported to 1.11.2
- Probably countless other fixes

+++ src/main/java/sonar/core/network/sync/SyncTagType.java
@@ -92,12 +92,12 @@ public abstract class SyncTagType<T> extends SyncPart {
 			this.current = (long) 0;
 		}
 
-		public void increaseBy(int i) {
+        public void increaseBy(long i) {
 			setObject(getObject() + i);
 			markChanged();
 		}
 
-		public void decreaseBy(int i) {
+        public void decreaseBy(long i) {
 			setObject(getObject() - i);
 			markChanged();
 		}
@@ -232,5 +232,4 @@ public abstract class SyncTagType<T> extends SyncPart {
 	public String toString() {
 		return current.toString();
 	}
-	
 }
